created helper functions file to generate array of dates
Working on query for teams and appending proper data to each entry.  The display table needs to have a range of every date in between the user selection, not just the selection range itself.  Otherwise there will be gaps.
