+++ src/services/FantasyServices.js
@@ -78,14 +78,15 @@ export async function getTeams() {
   return players
 }
 
-export async function getTeamSchedules(dateRange) {
+export async function getGamesWithinThis(dateRange, teamId) {
   let schedules
   try {
     schedules = axi_fantasy.get({
       endpoint: 'schedule',
       payload: {
         startDate: dateRange[0],
-        endDate: dateRange[1]
+        endDate: dateRange[1],
+        teamId: teamId
       }
     })
   }
